backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.0",
    "diffusers>=0.35.1",
    "dotenv>=0.9.9",
    "portia-sdk-python[all]>=0.7.0",
    "torch>=2.8.0",
    "transformers>=4.55.3",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
]
